refactor[FE]: Move Background component to  separete file
